requirements.txt
aiogram==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
bcrypt==4.1.1
